Add status field
