+++ packages/data/src/data.ts
@@ -1,22 +1,12 @@
 import type { Category, Auth, LinkLabel, API } from "./types";
 
+import { sortByName } from "./utilities";
+
 const AUTHs: Auth[] = ["No", "API Key", "OAuth"];
 
 const CATEGORIEs: Category[] = ["Location", "Other"];
 
 const LINKLABELs: LinkLabel[] = ["Website", "Github"];
-
-function sortByName(data: API[]): API[] {
-  return [
-    ...data.sort((a, b) => {
-      let x = a.name.toLowerCase();
-      let y = b.name.toLowerCase();
-      if (x > y) return 1;
-      if (x < y) return -1;
-      return 0;
-    }),
-  ];
-}
 
 const APIs: API[] = sortByName([
   {
